+++ predictionscorer/_common.py
@@ -1,6 +1,6 @@
 import math
 from decimal import Decimal
-from typing import Optional, Union
+from typing import Union
 
 
 def to_decimal(probability: Union[Decimal, float, int]) -> Decimal:
@@ -12,9 +12,9 @@ def to_decimal(probability: Union[Decimal, float, int]) -> Decimal:
     return Decimal(probability)
 
 
-def _inverse_probability(probability: Decimal) -> Decimal:
+def inverse_probability(probability: Decimal) -> Decimal:
     return Decimal(1) - probability
 
 
-def _log(value: Decimal) -> Decimal:
+def log(value: Decimal) -> Decimal:
     return Decimal(str(math.log2(value)))
